+++ TP01/src/com/g7/rules/DropSeven.java
@@ -24,7 +24,7 @@ public class DropSeven implements GPSRule{
 	}
 
 	@Override
-	public Integer getCost() {
+	public float getCost() {
 		return 1;
 	}
