+++ TP01/src/com/g7/rules/DropTen.java
@@ -24,7 +24,7 @@ public class DropTen implements GPSRule{
 	}
 
 	@Override
-	public Integer getCost() {
+	public float getCost() {
 		return 1;
 	}
